+++ src/rng.rs
@@ -1,9 +1,9 @@
 use fastrand;
 
-pub fn random() -> f32 {
-    fastrand::f32()
+pub fn random() -> f64 {
+    fastrand::f64()
 }
 
-pub fn random_in_range(min: f32, max: f32) -> f32 {
+pub fn random_in_range(min: f64, max: f64) -> f64 {
     min + (max - min) * random()
 }
